+++ plugin/index.js
@@ -2,14 +2,14 @@
 
 const { DOMAPIC } = require('../lib/statics')
 const SwitchFactory = require('../lib/plugins/SwitchFactory')
-const ButtonFactory = require('../lib/plugins/ButtonFactory')
+const ContactSensorFactory = require('../lib/plugins/ContactSensorFactory')
 
 module.exports = function (homebridge) {
   const Switch = new SwitchFactory(homebridge.hap.Service, homebridge.hap.Characteristic)
   homebridge.registerAccessory(DOMAPIC, Switch.name, Switch)
 
-  const Button = new ButtonFactory(homebridge.hap.Service, homebridge.hap.Characteristic)
-  homebridge.registerAccessory(DOMAPIC, Button.name, Button)
+  const ContactSensor = new ContactSensorFactory(homebridge.hap.Service, homebridge.hap.Characteristic)
+  homebridge.registerAccessory(DOMAPIC, ContactSensor.name, ContactSensor)
 
   // TODO, register accesories for all HomeKit accesory types
 }