+++ server.js
@@ -8,7 +8,7 @@ const { debounce } = require('lodash')
 const options = require('./lib/options')
 
 const abilitiesBridgeApi = require('./lib/api/abilities-bridge.json')
-const Abilities = require('./lib/Abilities')
+const Accessories = require('./lib/Accessories')
 const AbilitiesBridge = require('./lib/api/AbilitiesBridge')
 const Homebridge = require('./lib/Homebridge')
 const HomebridgeConfig = require('./lib/HomebridgeConfig')
@@ -17,20 +17,25 @@ domapic.createPlugin({
   packagePath: path.resolve(__dirname),
   customConfig: options
 }).then(async dmpcPlugin => {
-  const abilities = new Abilities(dmpcPlugin)
+  const accessories = new Accessories(dmpcPlugin)
   const abilitiesBridge = new AbilitiesBridge(dmpcPlugin)
   const homebridge = new Homebridge(dmpcPlugin)
   const homebridgeConfig = new HomebridgeConfig(dmpcPlugin)
 
   const restartHomebridge = debounce(async (event) => {
-    await homebridgeConfig.write(await abilities.get())
+    await homebridgeConfig.write(await accessories.get())
     homebridge.restart()
-  }, 5000)
+  }, 10000)
 
+  dmpcPlugin.events.on('service:updated', restartHomebridge)
+  dmpcPlugin.events.on('service:created', restartHomebridge)
+  dmpcPlugin.events.on('service:deleted', restartHomebridge)
+  dmpcPlugin.events.on('servicePluginConfig:created', restartHomebridge)
+  dmpcPlugin.events.on('servicePluginConfig:updated', restartHomebridge)
+  dmpcPlugin.events.on('servicePluginConfig:deleted', restartHomebridge)
   dmpcPlugin.events.on('ability:updated', restartHomebridge)
   dmpcPlugin.events.on('ability:created', restartHomebridge)
   dmpcPlugin.events.on('ability:deleted', restartHomebridge)
-  dmpcPlugin.events.on('service:updated', restartHomebridge)
   dmpcPlugin.events.once('connection', restartHomebridge)
 
   await dmpcPlugin.api.extendOpenApi(abilitiesBridgeApi)