+++ test/unit/plugin/index.specs.js
@@ -1,13 +1,13 @@
 const test = require('narval')
 
 const SwitchFactoryMocks = require('../lib/plugins/SwitchFactory.mocks')
-const ButtonFactoryMocks = require('../lib/plugins/ButtonFactory.mocks')
+const ContactSensorFactoryMocks = require('../lib/plugins/ContactSensorFactory.mocks')
 const HomebridgeMocks = require('../Homebridge.mocks')
 
 test.describe('plugin', () => {
   let sandbox
   let switchFactory
-  let buttonFactory
+  let contactSensorFactory
   let homebridge
   let plugin
 
@@ -15,7 +15,7 @@ test.describe('plugin', () => {
     sandbox = test.sinon.createSandbox()
     homebridge = new HomebridgeMocks()
     switchFactory = new SwitchFactoryMocks()
-    buttonFactory = new ButtonFactoryMocks()
+    contactSensorFactory = new ContactSensorFactoryMocks()
     plugin = require('../../../plugin/index')
     plugin(homebridge.stubs)
   })
@@ -23,7 +23,7 @@ test.describe('plugin', () => {
   test.after(() => {
     sandbox.restore()
     switchFactory.restore()
-    buttonFactory.restore()
+    contactSensorFactory.restore()
   })
 
   test.it('should create a new Switch accessory', () => {
@@ -34,11 +34,11 @@ test.describe('plugin', () => {
     test.expect(homebridge.stubs.registerAccessory.getCall(0).args[2]).to.equal(switchFactory.stubs.instance)
   })
 
-  test.it('should create a new Button accessory', () => {
-    test.expect(buttonFactory.stubs.Constructor).to.have.been.calledWith(homebridge.stubs.hap.Service, homebridge.stubs.hap.Characteristic)
+  test.it('should create a new ContactSensor accessory', () => {
+    test.expect(contactSensorFactory.stubs.Constructor).to.have.been.calledWith(homebridge.stubs.hap.Service, homebridge.stubs.hap.Characteristic)
   })
 
-  test.it('should register button accesory in homebridge', () => {
-    test.expect(homebridge.stubs.registerAccessory.getCall(1).args[2]).to.equal(buttonFactory.stubs.instance)
+  test.it('should register contactSensor accesory in homebridge', () => {
+    test.expect(homebridge.stubs.registerAccessory.getCall(1).args[2]).to.equal(contactSensorFactory.stubs.instance)
   })
 })